Use hot socket so worker does not have to wait for its own ThriftServer
Summary:
The entire premise behind `canConnectToMyself_` was that it is not clear how `BistroWorkerHandler::heartbeat()` can ascertain whether or not its `server_` can already accept a request from the Bistro scheduler.

There is probably some getter in `ThriftServer` that can be used to infer this, and so we could maybe busy-wait on that, but ... it appears that passing a "hot" (aka already listening / accepting) socket to the server works just fine.

This is nicer because it reduces the amount of code, and puts all the waiting in the kernel.

Differential Revision: D27519932

fbshipit-source-id: 9ab2f2d56b5d892b84c00ca7aa8b485e63fd2908